refactor: move CANONICAL_PERSON to a neutral constant
The STAY tournament-admin surfaces (registrations, declarations) imported
CANONICAL_PERSON from account/auth/hiveid.constants — the account tree that
lifts out to the HiveID IdP. Move the constant to a neutral
src/common/constants/canonicalPerson so those surfaces don't reach into the
account tree; re-export from hiveid.constants for the MOVE-side callers.

A Phase-3 re-parenting prerequisite (ACCOUNT_MOVE_PHASE2_3_PLAN.md §2c):
behavior-preserving, no runtime change. Full affected suites green; account
boundary clean.
